requirements.txt
streamlit
google-generativeai
pandas
numpy
plotly